Move SpillCipher to SPI
