+++ presto-main/src/main/java/com/facebook/presto/spiller/FileSingleStreamSpiller.java
@@ -20,6 +20,7 @@ import com.facebook.presto.operator.SpillContext;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.page.SerializedPage;
+import com.facebook.presto.spi.spiller.SpillCipher;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.ImmutableList;
