+++ presto-main/src/test/java/com/facebook/presto/execution/buffer/TestingPagesSerdeFactory.java
@@ -17,8 +17,8 @@ import com.facebook.presto.block.BlockEncodingManager;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.block.BlockEncodingSerde;
 import com.facebook.presto.spi.page.SerializedPage;
+import com.facebook.presto.spi.spiller.SpillCipher;
 import com.facebook.presto.spi.type.TestingTypeManager;
-import com.facebook.presto.spiller.SpillCipher;
 import io.airlift.compress.Compressor;
 import io.airlift.compress.Decompressor;
 import io.airlift.compress.lz4.Lz4Compressor;
