+++ presto-main/src/test/java/com/facebook/presto/spiller/TestSpillCipherPagesSerde.java
@@ -19,6 +19,7 @@ import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.page.SerializedPage;
+import com.facebook.presto.spi.spiller.SpillCipher;
 import com.facebook.presto.spi.type.Type;
 import com.google.common.collect.ImmutableList;
 import org.testng.Assert.ThrowingRunnable;
